feat: move components from shared to employees
