+++ cv-gen/src/app/modules/employees/components/level-input/level-input.component.scss
