+++ cv-gen/src/app/modules/employees/components/skills/skills.component.scss
