+++ cv-gen/src/app/modules/employees/components/skills/skills.component.ts
@@ -13,7 +13,7 @@ import {
   FormGroupDirective,
 } from '@angular/forms';
 import { SkillsInterface } from '../../../core/models/skills.model';
-import { bothFieldsRequired } from '../../../core/helpers/skill.validator';
+import { bothFieldsRequired } from '../../helpers/skill.validator';
 
 @Component({
   selector: 'app-skills',
