+++ cv-gen/src/app/modules/employees/helpers/skill.validator.ts
@@ -1,5 +1,5 @@
 import { AbstractControl, ValidationErrors, ValidatorFn } from '@angular/forms';
-import { SkillsInterface } from '../models/skills.model';
+import { SkillsInterface } from '../../core/models/skills.model';
 
 export function bothFieldsRequired(): ValidatorFn {
   return function (
